Change deployment type in friendlycode's HTML based on settings.DEV.
If settings.DEV is true, friendlycode will think the deployment type
is 'development'. Otherwise, it will think the deployment type is
'production'.

Note that if the app is being served from staging infrastructure,
the deployment type will be classified as 'production', since settings.DEV
is False. We might want to change this in the future... At the moment,
because staging is so similar to production, I don't know of a way to
detect whether we're a staging instance from Django code.

This should fix #201.