Add caching `api.views.get_page`.
Issue: #313

I added the `@cache_page` decorator which uses django's caching backend
to memoize the function. It also adds a 'Cache-Control: max-age`
directive to the response.

I also added in an `ETag` header (generated from the sha1 hash of the
html) and a 'Cache-Control: public'  directive to allow for conditional
GET requests and proxy caching.