Change caching strategy for `api.views.get_page`.
[Issue #313]

We are going to use a short-lived cache for now, one day instead of one
year. We'll still use the `@cache_page` decorator because if server-side
caching backends ever get enabled and defined in settings.py, it will be
able to take advantage of them automatically.

This changeset also adds a `SECURITY_VERSION` to the sanitize module
which is included in the used in the`generate_etag` function to make
ensure that whenever we update the sanitization algorithm all previous
ETags get invalidated. Because we have `Cache-Control: max-age=86400`,
browsers with a primed cache will still see old data for up to 24 hours,
but a future conditional GET will require a new full response.