Don't sanitize style attributes.
This should fix #32. However, it does so in a kind of weird way, because,
after reading bleach's source code, it appears that style attributes
MUST be sanitized via whitelisted CSS properties in bleach--there's no
easy way to tell bleach to allow style attributes with any arbitrary
content in them. So this fix works by monkeypatching bleach.BleachSanitizer,
which feels icky. But hey, Playdoh seems to do this kind of thing when
it absolutely needs to, so...

Also, I added a TODO mentioning that we should eventually upstream a
real fix to bleach, so we can get rid of this monkeypatch.