Better rebasing of static URLs for learning projects.
Previously, we'd do a find/replace of 'static/', which meant that
URLs in learning projects like http://foo.com/static/blah.png got
converted to e.g. http://foo.com/http://thimble.org/static/blah.png.
Now we use a regexp with a negative lookbehind assertion to ensure that
we only rebase static URLs when they're not preceded with a '/'.