Added utf8 charset/collation params to CREATE DATABASE example.
Extremely frustrating. Was running into a bug in development that
I couldn't reproduce in a test because Django takes care of
creating the test DB with the requisite utf8 params, but there
are no checks at startup to ensure that the production DB has
the same params.

Ideally we should probably add such checks at startup.