Support real clopenbadger instances.
I wanted to keep it easy for us to use fake clopenbadger jquery middleware
for now, and at the moment this means using if statements in jinja2. I
don't like this because it involves dynamic generation of JS, but I can't
think of an easier way of doing this right now.